Split roc indicator into several files.
