+++ js/Stock/Indicators/ROC/ROCIndicator.js
@@ -21,9 +21,9 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../../Core/Series/Series.js';
+import BaseSeries from '../../../Core/Series/Series.js';
 var SMAIndicator = BaseSeries.seriesTypes.sma;
-import U from '../../Core/Utilities.js';
+import U from '../../../Core/Utilities.js';
 var isArray = U.isArray, merge = U.merge, extend = U.extend;
 /* eslint-disable require-jsdoc */
 // Utils:
