+++ ts/Stock/Indicators/ROC/ROCOptions.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type {
+    SMAOptions,
+    SMAParamsOptions
+} from '../SMA/SMAOptions';
+
+/* *
+*
+*  Declarations
+*
+* */
+
+export interface ROCOptions extends SMAOptions {
+    params?: ROCParamsOptions;
+}
+
+export interface ROCParamsOptions extends SMAParamsOptions {
+    // for inheritance
+}
+
+export default EMAOptions;
