+++ ts/Stock/Indicators/ROC/ROCPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ROCIndicator from './ROCIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class ROCPoint extends SMAPoint {
+    public series: ROCIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default ROCPoint;
